requirements.txt
Django==3.1.4
djangorestframework==3.12.2
django-rest-swagger==2.2.0
coreapi==2.3.3
uwsgi==2.0.19.1
mysqlclient==2.0.1
django-tables2==2.3.1
